[TASK] Read sitemaps from robots.txt
On top level or with robots.txt as path in the url, the sitecrawler will check this file to gather all sitemaps defined in there. This allows the easy crawling of a multilingual TYPO3 project.

The testing process must be changed to allow to test this feature. The GeneralUtility::getUrl() Method needs to be mocked and the sitemap/robots.txt data provided for tests.